refactor: MAX_LENGTH 위치 변경
