+++ client/src/features/Organization/New/constants/validationRules.ts
@@ -1,0 +1,4 @@
+export const MAX_LENGTH = {
+  NAME: 30,
+  DESCRIPTION: 30,
+} as const;
